Better background
There is now support for background selection from radii around each
occurrence record. This means that, for conditional likelihood
estimation, the background is a better sample of “available climate” if
tuned appropriately.

Second set of edits allows user defined intervals (range, 95, 99, and
99.9% confidence intervals) to be used to “clip” the PDF distributions.
This limits the possible estimated climate to the range of overlap of
all taxa which may be sensitive to poor sampling (thus the 99.9%CI
option as a conservative approach).